feat(unic-th): split theme + make more consistent
