+++ themes/theme-unicorn/src/index.ts
@@ -1,193 +1,7 @@
 import { type Theme } from '@marigold/components';
 import { colors } from './colors';
-
-const selectButton = {
-  appearance: 'none',
-  position: 'relative',
-  display: 'inline-flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  overflow: 'hidden',
-  outline: 'none',
-  width: '100%',
-  lineHeight: '32px',
-  px: 'xsmall',
-  color: 'inherit',
-  bg: 'transparent',
-} as const;
-const button = {
-  root: {
-    position: 'relative',
-    fontFamily: 'body',
-    fontSize: 'xsmall',
-    fontWeight: 'body',
-    border: 'none',
-    borderRadius: '8px',
-    display: 'inline-flex',
-  },
-  primary: {
-    color: 'background',
-    bg: 'primary',
-    ':hover': {
-      color: 'background',
-      bg: colors.gray20,
-      cursor: 'pointer',
-    },
-    ':disabled': {
-      color: colors.gray40,
-      bg: colors.gray20,
-      cursor: 'not-allowed',
-    },
-  },
-  secondary: {
-    color: 'background',
-    bg: 'secondary',
-    ':hover': {
-      color: 'background',
-      bg: colors.gray60,
-      cursor: 'pointer',
-    },
-    ':disabled': {
-      color: colors.gray40,
-      bg: colors.gray20,
-      cursor: 'not-allowed',
-    },
-  },
-  ghost: {
-    color: 'secondary',
-    border: '1px solid',
-    outlineColor: colors.gray70,
-    ':hover': {
-      color: 'secondary',
-      bg: colors.gray30,
-      cursor: 'pointer',
-    },
-    ':disabled': {
-      color: 'disabled',
-      bg: colors.gray00,
-      outlineColor: 'disabled',
-      cursor: 'not-allowed',
-    },
-  },
-  text: {
-    color: 'secondary',
-    ':hover': {
-      color: 'secondary',
-      outlineColor: colors.gray70,
-      bg: colors.gray30,
-      cursor: 'pointer',
-    },
-    ':disabled': {
-      color: 'disabled',
-      bg: colors.gray00,
-      cursor: 'not-allowed',
-      outline: 'none',
-    },
-  },
-  menu: {
-    color: 'secondary',
-    bg: 'background',
-    ':hover': {
-      color: 'background',
-      bg: colors.gray60,
-      cursor: 'pointer',
-    },
-    ':disabled': {
-      color: colors.gray40,
-      bg: colors.gray20,
-      cursor: 'not-allowed',
-    },
-  },
-  select: {
-    ...selectButton,
-    border: '1px solid transparent',
-    borderColor: colors.gray40,
-    ':hover': {
-      cursor: 'pointer',
-    },
-    ':focus': {
-      boxShadow: '0 0 0 2px ' + colors.blue60,
-      border: '1px solid transparent',
-    },
-    ':disabled': {
-      bg: colors.gray20,
-      color: 'disabled',
-      cursor: 'not-allowed',
-    },
-    error: {
-      ...selectButton,
-      border: '1px solid',
-      borderRadius: '8px',
-      borderColor: 'error',
-    },
-    open: {
-      ...selectButton,
-      border: '1px solid transparent',
-      borderTopColor: colors.gray40,
-      borderLeftColor: colors.gray40,
-      borderRightColor: colors.gray40,
-      borderTopRightRadius: '8px',
-      borderTopLeftRadius: '8px',
-    },
-    errorOpened: {
-      ...selectButton,
-      border: '1px solid transparent',
-      borderTopColor: 'error',
-      borderLeftColor: 'error',
-      borderRightColor: 'error',
-      borderTopRightRadius: '8px',
-      borderTopLeftRadius: '8px',
-    },
-  },
-  close: {
-    color: 'text',
-    bg: 'transparent',
-    ':hover': {
-      cursor: 'pointer',
-    },
-  },
-} as const;
-const text = {
-  root: {
-    fontFamily: 'body',
-    fontSize: 'xsmall',
-    lineHeight: 'body',
-    fontWeight: 'body',
-  },
-  span: {
-    display: 'inline-block',
-  },
-} as const;
-const selectOption = {
-  fontFamily: 'body',
-  fontSize: 'xsmall',
-  fontWeight: 'body',
-  lineHeight: '32px',
-  outline: 'none',
-  cursor: 'pointer',
-  color: 'text',
-  px: 'xsmall',
-  listStyle: 'none',
-} as const;
-const sliderThumb = {
-  WebkitAppearance: 'none',
-  boxSizing: 'border-box',
-  border: '4px solid ' + colors.gray70,
-  width: '16px',
-  height: '16px',
-  background: colors.gray00,
-  borderRadius: '8px',
-  cursor: 'pointer',
-  marginTop: '-4px',
-} as const;
-const sliderTrack = {
-  WebkitAppearance: 'none',
-  width: '100%',
-  height: '8px',
-  background: colors.gray30,
-  borderRadius: '8px',
-  border: 'none',
-} as const;
+import { components } from './components';
+import { typography } from './typography';
 
 const theme: Theme = {
   space: {
@@ -200,31 +14,10 @@ const theme: Theme = {
     xlarge: 42,
     xxlarge: 64,
   },
-  fonts: {
-    body: 'Arial',
-    heading: 'Roboto',
-  },
-  fontSizes: {
-    xxsmall: '0.66rem',
-    xsmall: '1rem',
-    small: '1.33rem',
-    medium: '1.66rem',
-    large: '2rem',
-    xlarge: '2.33rem',
-  },
-  fontWeights: {
-    body: 300,
-    heading: 800,
-    bold: 600,
-  },
-  lineHeights: {
-    body: 2,
-    heading: 1.5,
-  },
   colors: {
     ...colors,
-    text: 'gray90',
-    background: 'gray00',
+    text: colors.gray90,
+    background: colors.gray00,
     primary: '#673ab7',
     secondary: '#9575cd',
     disabled: '#e9e7eb',
@@ -240,540 +33,14 @@ const theme: Theme = {
       fontFamily: 'body',
     },
   },
-  alert: {
-    error: {
-      alignItems: 'center',
-      borderStyle: 'solid',
-      borderColor: 'error',
-      borderWidth: '3px 3px 3px 0px',
-    },
-    warning: {
-      alignItems: 'center',
-      borderStyle: 'solid',
-      borderColor: 'warning',
-      borderWidth: '3px 3px 3px 0px',
-    },
-    success: {
-      alignItems: 'center',
-      borderStyle: 'solid',
-      borderColor: 'success',
-      borderWidth: '3px 3px 3px 0px',
-    },
+  radii: {
+    none: 0,
+    small: 2,
+    medium: 4,
+    large: 8,
   },
-  badge: {
-    default: {
-      display: 'inline-flex',
-      alignItems: 'center',
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      borderRadius: '10px',
-      border: '2px solid transparent',
-      whiteSpace: 'nowrap',
-      padding: '0.25rem 0.75rem',
-      mx: '0.5rem',
-    },
-  },
-  button: {
-    large: {
-      lineHeight: '46px',
-      paddingX: 'xlarge',
-    },
-    small: {
-      lineHeight: '30px',
-      paddingX: 'medium',
-    },
-    xsmall: {
-      lineHeight: '16px',
-    },
-    primary: {
-      ...button.root,
-      ...button.primary,
-    },
-    secondary: {
-      ...button.root,
-      ...button.secondary,
-    },
-    ghost: {
-      ...button.root,
-      ...button.ghost,
-    },
-    text: {
-      ...button.root,
-      ...button.text,
-    },
-    menu: {
-      ...button.root,
-      ...button.menu,
-    },
-    select: {
-      ...button.root,
-      ...button.select,
-    },
-    close: {
-      ...button.root,
-      ...button.close,
-    },
-  },
-  card: {
-    default: {
-      maxWidth: '500px',
-      background: colors.gray00,
-      p: 'small',
-      boxShadow: '8px 8px 4px rgba(201, 177, 255, 0.5)',
-      borderRadius: '8px',
-    },
-  },
-  checkbox: {
-    default: {
-      ariaHidden: 'true',
-      mr: 2,
-      verticalAlign: 'middle',
-      ':hover': { cursor: 'pointer' },
-      'input:disabled ~ &': {
-        cursor: 'not-allowed',
-      },
-    },
-    checked: {
-      fill: 'primary',
-      stroke: '#311b92',
-      disabled: {
-        fill: colors.gray30,
-        stroke: colors.gray30,
-      },
-      icon: {
-        fill: colors.gray00,
-      },
-    },
-    unchecked: {
-      fill: colors.gray00,
-      stroke: colors.gray40,
-      disabled: {
-        stroke: colors.gray30,
-      },
-      error: {
-        stroke: 'error',
-      },
-    },
-  },
-  radio: {
-    default: {
-      ariaHidden: 'true',
-      mr: 2,
-      verticalAlign: 'middle',
-      ':hover': { cursor: 'pointer' },
-      'input:disabled ~ &': {
-        color: 'muted',
-        cursor: 'not-allowed',
-      },
-    },
-    checked: {
-      fill: 'primary',
-      stroke: '#311b92',
-      disabled: {
-        fill: colors.gray30,
-        stroke: colors.gray30,
-      },
-      circle: {
-        fill: colors.gray00,
-      },
-    },
-    unchecked: {
-      fill: colors.gray00,
-      stroke: colors.gray40,
-      disabled: {
-        stroke: colors.gray30,
-      },
-      error: {
-        stroke: 'error',
-      },
-    },
-  },
-  dialog: {
-    wrapper: {
-      display: 'flex',
-      justifyContent: 'space-between',
-      borderRadius: '2px',
-      paddingLeft: 'large',
-      paddingBottom: 'large',
-    },
-    body: {
-      paddingTop: 'medium',
-    },
-    onClose: {
-      display: 'flex',
-      justifyContent: 'flex-end',
-      alignItems: 'start',
-      paddingTop: 'xsmall',
-      paddingX: 'xsmall',
-    },
-    modalWrapper: {
-      position: 'fixed',
-      zIndex: 100,
-      top: 0,
-      left: 0,
-      bottom: 0,
-      right: 0,
-      background: 'rgba(0, 0, 0, 0.5)',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    modalBody: {
-      background: '#ffffff',
-      minWidth: '510px',
-      minHeight: '240px',
-    },
-  },
-  divider: {
-    regular: {
-      m: 'none',
-      my: 'small',
-      border: 0,
-      borderBottom: '1px solid',
-    },
-    bold: {
-      m: 'none',
-      my: 'small',
-      border: 0,
-      borderBottom: '3px solid',
-    },
-    section: {
-      m: 'none',
-      mb: 'small',
-      border: 0,
-      borderBottom: '1px solid',
-      borderColor: colors.gray50,
-    },
-  },
-  image: {
-    fullWidth: {
-      maxWidth: '100%',
-      height: 'auto',
-    },
-  },
-  input: {
-    default: {
-      display: 'block',
-      fontFamily: 'body',
-      color: 'text',
-      border: 'none',
-      borderRadius: '2px',
-      boxShadow: '0 0 0 1px ' + colors.gray40,
-      outline: 'none',
-      padding: '0 12px',
-      lineHeight: '32px',
-      ':focus': {
-        boxShadow: '0 0 0 2px ' + colors.blue60,
-      },
-      ':disabled': {
-        bg: colors.gray20,
-        color: colors.gray40,
-        cursor: 'not-allowed',
-      },
-    },
-    error: {
-      fontFamily: 'body',
-      color: 'text',
-      border: 'none',
-      borderRadius: '2px',
-      boxShadow: '0 0 0 1px ' + colors.red60,
-      outline: 'none',
-      padding: '0 8px',
-      lineHeight: '32px',
-    },
-  },
-  label: {
-    above: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      lineHeight: 'body',
-      color: 'text',
-    },
-    inline: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      lineHeight: '2rem',
-      color: 'text',
-      display: 'inline-flex',
-      alignItems: 'center',
-    },
-    disabled: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      lineHeight: 'body',
-      color: colors.gray30,
-    },
-    section: {
-      fontFamily: 'body',
-      fontSize: 'small',
-      fontWeight: 'body',
-      lineHeight: '32px',
-      color: colors.gray50,
-    },
-  },
-  menuItem: {
-    default: {
-      display: 'block',
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      padding: 'xsmall',
-      bg: 'primary',
-      color: 'text',
-      ':hover': {
-        bg: 'secondary',
-        cursor: 'pointer',
-      },
-    },
-  },
-  message: {
-    warning: {
-      borderStyle: 'solid',
-      borderColor: 'warning',
-      borderWidth: '2px 2px 2px 16px',
-      padding: '8px 16px 16px',
-      color: 'warning',
-    },
-    error: {
-      borderStyle: 'solid',
-      borderColor: 'error',
-      borderWidth: '2px 2px 2px 16px',
-      padding: '8px 16px 16px',
-      color: 'error',
-    },
-    info: {
-      borderStyle: 'solid',
-      borderColor: 'info',
-      borderWidth: '2px 2px 2px 16px',
-      padding: '8px 16px 16px',
-      color: 'info',
-    },
-    title: {
-      mb: 'small',
-    },
-  },
-  text: {
-    body: {
-      ...text.root,
-      ...text.span,
-    },
-    headline1: {
-      ...text.root,
-      fontWeight: 'heading',
-      fontSize: 'xlarge',
-      marginBottom: 'medium',
-      '&:first-child': {
-        marginTop: 'none',
-      },
-    },
-    headline2: {
-      ...text.root,
-      fontWeight: 'heading',
-      fontSize: 'large',
-      marginTop: 'medium',
-      marginBottom: 'xsmall',
-      '&:first-child': {
-        marginTop: 'none',
-      },
-    },
-    headline3: {
-      ...text.root,
-      fontWeight: 'heading',
-      fontSize: 'medium',
-      marginTop: 'small',
-      marginBottom: 'small',
-      '&:first-child': {
-        marginTop: 'none',
-      },
-    },
-    headline4: {
-      ...text.root,
-      fontWeight: 'heading',
-      fontSize: 'small',
-    },
-    headline5: {
-      ...text.root,
-      fontWeight: 'heading',
-      fontSize: 'xsmall',
-    },
-    headline6: {
-      ...text.root,
-      fontSize: 'xsmall',
-      textTransform: 'uppercase',
-    },
-    link: {
-      color: 'primary',
-      ':hover': {
-        textDecoration: 'none',
-      },
-    },
-    menuItemLink: {
-      color: 'text',
-      textDecoration: 'none',
-    },
-  },
-  textarea: {
-    default: {
-      fontFamily: 'body',
-      lineHeight: '24px',
-      padding: '4px 8px',
-      color: 'text',
-      border: 0,
-      borderRadius: '8px',
-      outline: '2px solid',
-      ':focus': {
-        outline: '4px solid',
-        outlineColor: 'primary',
-      },
-      ':disabled': {
-        bg: colors.gray20,
-        color: colors.gray40,
-      },
-    },
-  },
-  select: {
-    __default: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 400,
-      lineHeight: '32px',
-      color: 'text',
-    },
-    disabled: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 400,
-      lineHeight: '32px',
-      color: 'disabled',
-      cursor: 'not-allowed',
-    },
-    listbox: {
-      __default: {
-        background: colors.gray00,
-        border: '1px solid transparent',
-        borderLeftColor: colors.gray40,
-        borderRightColor: colors.gray40,
-        borderBottomColor: colors.gray40,
-        borderBottomLeftRadius: '8px',
-        borderBottomRightRadius: '8px',
-      },
-      error: {
-        background: colors.gray00,
-        border: '1px solid transparent',
-        borderLeftColor: 'error',
-        borderRightColor: 'error',
-        borderBottomColor: 'error',
-        borderBottomRightRadius: '8px',
-        borderBottomLeftRadius: '8px',
-      },
-    },
-    section: {
-      fontFamily: 'body',
-      fontSize: 'xsmall',
-      fontWeight: 'body',
-      lineHeight: '32px',
-      px: 'xxsmall',
-      color: colors.gray50,
-    },
-    option: {
-      __default: {
-        ...selectOption,
-        ':focus': {
-          bg: colors.blue20,
-        },
-      },
-      selected: {
-        ...selectOption,
-        color: colors.gray00,
-        bg: colors.blue60,
-      },
-      disabled: {
-        ...selectOption,
-        color: colors.gray40,
-      },
-    },
-  },
-  slider: {
-    default: {
-      // styles need to be applied to range inputs in all browsers to override their basic appearance.
-      WebkitAppearance: 'none',
-      background: 'transparent',
-      borderColor: 'transparent',
-      color: 'transparent',
-      ':focus': {
-        outline: 'none',
-      },
-      // chrome, safari, opera (theres actually no webkit option to style the progress bar like in firefox)
-      '&::-webkit-slider-thumb': {
-        ...sliderThumb,
-      },
-      '&:focus::-webkit-slider-thumb': {
-        ...sliderThumb,
-        border: '4px solid',
-        borderColor: 'primary',
-      },
-      '&:disabled::-webkit-slider-thumb': {
-        ...sliderThumb,
-        border: '4px solid ' + colors.gray40,
-        background: colors.gray40,
-      },
-      '&::-webkit-slider-runnable-track': {
-        ...sliderTrack,
-      },
-      '&:focus::-webkit-slider-runnable-track': {
-        ...sliderTrack,
-        backgroundColor: 'primary',
-      },
-      '&:disabled::-webkit-slider-runnable-track': {
-        ...sliderTrack,
-        background: colors.gray40,
-      },
-      '&::-webkit-progress-value': {
-        background: colors.green60,
-      },
-      // firefox
-      '&::-moz-range-thumb': {
-        ...sliderThumb,
-      },
-      '&:focus::-moz-range-thumb': {
-        ...sliderThumb,
-        border: '4px solid',
-        borderColor: 'primary',
-      },
-      '&:disabled::-moz-range-thumb': {
-        ...sliderThumb,
-        border: '4px solid ' + colors.gray40,
-        background: colors.gray40,
-      },
-      '&::-moz-range-track': {
-        ...sliderTrack,
-      },
-      '&::-moz-range-progress': {
-        ...sliderTrack,
-        background: colors.gray70,
-      },
-      '&:focus::-moz-range-progress': {
-        ...sliderTrack,
-        backgroundColor: 'primary',
-      },
-      '&:disabled::-moz-range-progress': {
-        ...sliderTrack,
-        background: colors.gray40,
-      },
-    },
-  },
-  validation: {
-    error: {
-      ...text.root,
-      fontSize: 'xsmall',
-      textTransform: 'uppercase',
-      color: 'error',
-    },
-  },
+  ...typography,
+  ...components,
 } as const;
 
 export default theme;
